Deciphering-Hate@ACL/requirements.txt
beautifulsoup4==4.12.3
matplotlib==3.7.1
matplotlib-inline==0.1.7
nltk==3.8.1
numpy==1.26.4
pandas==2.1.4
Pillow==9.4.0
scikit-learn==1.3.2
seaborn==0.13.1
sklearn-pandas==2.2.0
torch==2.4.0
torchsummary==1.5.1
torchvision==0.19.0
tqdm==4.66.5
transformers==4.44.2
git+https://github.com/openai/CLIP.git
tensorflow==2.17.0
tensorboard==2.17.0
sentencepiece==0.1.99
gdown
Madgrad